+++ docker/postgis/initdb.d/40_init.sql
@@ -1,0 +1,4 @@
+-- 40_init.sql --- Enable OMS's WFS to use hash store for Attributes
+
+\connect omsb_db
+CREATE EXTENSION hstore;