+++ src/oms_sensemaking/inference/rules/incursions.py
@@ -1,7 +1,6 @@
 
 from oms_sdk.generated.generated_graphql_client import (
     ActivityQuery,
-    ActivityState,
     AttributeQuery,
     AttributesAttributesData,
     AttributeType,
@@ -171,7 +170,7 @@ class Incursion(BaseRule):
             classIri=SETTINGS.inference_incursion_class_iri,
             name="Incursion",
             description=f"Incursion detected into {geo_of_interest}",  # edit based on actual geo of interests format
-            state=ActivityState.UNKNOWN,
+            state=SETTINGS.inference_incursion_activity_state,
             nodeId=observation.nodeId,
             observationIds=[observation.id],
             startTime=observation.startTime,