+++ tests_int/mil_symbol/test_mil_symbol.py
@@ -1,6 +1,7 @@
 """MilSymbol Sensemaker Integration Tests"""
-from typing import List
+from typing import List, Optional
 from unittest import mock
+from uuid import uuid4
 
 from oms_sdk import DEFAULT_ACM
 from oms_sdk.generated.generated_graphql_client import (
@@ -11,6 +12,8 @@ from oms_sdk.generated.generated_graphql_client import (
     CreateNodeInput,
     NodeNode,
     ObjectTier,
+    RelationshipRelationship,
+    SourceSource,
     UpdateNodeInput,
 )
 from sqlalchemy import delete, select
@@ -22,7 +25,6 @@ from oms_sensemaking.models.sensemaking import Finding, FindingType
 
 
 def create_node(oms_crud_tool: OmsCrudTool, class_iri: str, symbol_id_code: str) -> NodeNode:
-
     create_node_input = CreateNodeInput(
         acm=DEFAULT_ACM,
         name="test node",
@@ -40,22 +42,25 @@ def create_attribute(
         oms_node: NodeNode,
         attribute_iri: str,
         attribute_value: str,
-        mock_source) -> AttributeAttribute:
+        mock_source: SourceSource,
+        tags: Optional[List[str]] = None
+        ) -> AttributeAttribute:
 
     attr = AttributeAttribute.model_construct(
+        id=uuid4(),
         attributeType=AttributeType.STRING,
         attributeValue=attribute_value,
         attributeIri=attribute_iri,
         acm=DEFAULT_ACM,
         sourceId=mock_source.id,
         confidence=Confidence.HIGH,
-        nodeId=oms_node.id
+        nodeId=oms_node.id,
+        tags=tags
     )
     return attr
 
 
 def test_execute(mock_source, db, mil_symbol_rules):
-
     oms_crud_tool = OmsCrudTool()
     sensemaker = MilSymbolSensemaker(mil_symbol_rules, oms_crud_tool)
 
@@ -102,6 +107,7 @@ def test_execute(mock_source, db, mil_symbol_rules):
                 attributeIri=SETTINGS.mil_symbol_settings.symbol_attribute_iri,
                 attributeType=AttributeType.STRING,
                 attributeValue=symbol.new_symbol_id_code,
+                attributeName="Icon",
                 confidence=Confidence.HIGH.value,
                 acm=symbol.acm,
                 nodeId=oms_node.id,
@@ -111,7 +117,6 @@ def test_execute(mock_source, db, mil_symbol_rules):
 
     # clear findings
     findings = db.execute(delete(Finding))
-
 
     # case 2
     oms_node = create_node(
@@ -156,6 +161,7 @@ def test_execute(mock_source, db, mil_symbol_rules):
                 attributeIri=SETTINGS.mil_symbol_settings.symbol_attribute_iri,
                 attributeType=AttributeType.STRING,
                 attributeValue=symbol.new_symbol_id_code,
+                attributeName="Icon",
                 confidence=Confidence.HIGH.value,
                 acm=symbol.acm,
                 nodeId=oms_node.id,
@@ -206,6 +212,7 @@ def test_execute(mock_source, db, mil_symbol_rules):
                 attributeIri=SETTINGS.mil_symbol_settings.symbol_attribute_iri,
                 attributeType=AttributeType.STRING,
                 attributeValue=symbol.new_symbol_id_code,
+                attributeName="Icon",
                 confidence=Confidence.HIGH.value,
                 acm=symbol.acm,
                 nodeId=oms_node.id,
@@ -214,4 +221,139 @@ def test_execute(mock_source, db, mil_symbol_rules):
         )
 
     # clear findings
+    findings = db.execute(delete(Finding))
+
+    # case 4 (Attribute Update)
+    oms_node = create_node(
+        oms_crud_tool,
+        "http://www.ontologyrepository.com/CommonCoreOntologies/Spacecraft",
+        "10-0-0-01-0-0-00-000000-00-00"
+    )
+
+    oms_attribute = create_attribute(
+        oms_node, SETTINGS.mil_symbol_settings.affiliation_iris[0], "neutral", mock_source)
+
+    sensemaker.get_context = mock.MagicMock(return_value=create_attribute(
+        oms_node, "https://foundry.ai.mil/MIDB_GST/v1/Target_Vetted", "true", mock_source))
+    sensemaker.get_affiliation = mock.MagicMock(return_value=oms_attribute)
+    sensemaker.get_status = mock.MagicMock(return_value=create_attribute(
+        oms_node, SETTINGS.mil_symbol_settings.status_iris[0], "present", mock_source))
+    sensemaker.get_node_ancestors_iris = mock.MagicMock(
+        return_value=["http://www.ontologyrepository.com/CommonCoreOntologies/Vehicle"])
+
+    # test an Attribute being executed by the Sensemaker
+    symbols: List[SymbolCodeUpdate] = sensemaker.execute(oms_attribute)
+    assert len(symbols) == 2
+
+    oms_crud_tool.update_node.assert_any_call(
+            UpdateNodeInput(
+                id=oms_node.id,
+                symbolIdCode=symbols[1].new_symbol_id_code
+            )
+        )
+
+    # check that symbols exist in Findings table
+    findings = db.execute(
+        select(Finding).filter(Finding.finding_type == FindingType.MIL_SYMBOL_UPDATE.value)).scalars().all()
+
+    assert len(findings) == 2
+    assert findings[0].finding_data["new_symbol_id_code"] == "10-0-4-05-0-0-00-000000-00-00"
+    assert findings[1].finding_data["new_symbol_id_code"] == "SNPP------*****"
+
+    for symbol in symbols:
+        oms_crud_tool.create_attribute.assert_any_call(
+            CreateAttributeInput(
+                tags=SETTINGS.mil_symbol_settings.mil_symbol_sensemaker_tags,
+                attributeIri=SETTINGS.mil_symbol_settings.symbol_attribute_iri,
+                attributeType=AttributeType.STRING,
+                attributeValue=symbol.new_symbol_id_code,
+                attributeName="Icon",
+                confidence=Confidence.HIGH.value,
+                acm=symbol.acm,
+                nodeId=oms_node.id,
+                sourceId=mock_source.id
+            )
+        )
+
+    # clear findings
+    findings = db.execute(delete(Finding))
+
+    # case 5 (Ignore due to Tags from this Sensemaker)
+    oms_node = create_node(
+        oms_crud_tool,
+        "http://www.ontologyrepository.com/CommonCoreOntologies/Spacecraft",
+        "10-0-0-01-0-0-00-000000-00-00"
+    )
+
+    oms_attribute = create_attribute(
+        oms_node,
+        SETTINGS.mil_symbol_settings.affiliation_iris[0],
+        "neutral",
+        mock_source,
+        # The sensemaker should ignore Attribute creates/updates it wrote
+        tags=SETTINGS.mil_symbol_settings.mil_symbol_sensemaker_tags)
+
+    sensemaker.get_context = mock.MagicMock(return_value=create_attribute(
+        oms_node, "https://foundry.ai.mil/MIDB_GST/v1/Target_Vetted", "true", mock_source))
+    sensemaker.get_affiliation = mock.MagicMock(return_value=oms_attribute)
+    sensemaker.get_status = mock.MagicMock(return_value=create_attribute(
+        oms_node, SETTINGS.mil_symbol_settings.status_iris[0], "present", mock_source))
+    sensemaker.get_node_ancestors_iris = mock.MagicMock(
+        return_value=["http://www.ontologyrepository.com/CommonCoreOntologies/Vehicle"])
+
+    # test an Attribute being executed by the Sensemaker
+    symbols: List[SymbolCodeUpdate] = sensemaker.execute(oms_attribute)
+    assert len(symbols) == 0
+
+    # check that symbols exist in Findings table
+    findings = db.execute(
+        select(Finding).filter(Finding.finding_type == FindingType.MIL_SYMBOL_UPDATE.value)).scalars().all()
+
+    assert len(findings) == 0
+
+    # case 6 (Ignore 'Icon' Attributes)
+    oms_node = create_node(
+        oms_crud_tool,
+        "http://www.ontologyrepository.com/CommonCoreOntologies/Spacecraft",
+        "10-0-0-01-0-0-00-000000-00-00"
+    )
+
+    # ICON's are published by this Sensemaker so it doesn't need to process them
+    oms_attribute = create_attribute(
+        oms_node, SETTINGS.mil_symbol_settings.symbol_attribute_iri, "SNPP------*****", mock_source)
+
+    sensemaker.get_context = mock.MagicMock(return_value=create_attribute(
+        oms_node, "https://foundry.ai.mil/MIDB_GST/v1/Target_Vetted", "true", mock_source))
+    sensemaker.get_affiliation = mock.MagicMock(return_value=oms_attribute)
+    sensemaker.get_status = mock.MagicMock(return_value=create_attribute(
+        oms_node, SETTINGS.mil_symbol_settings.status_iris[0], "present", mock_source))
+    sensemaker.get_node_ancestors_iris = mock.MagicMock(
+        return_value=["http://www.ontologyrepository.com/CommonCoreOntologies/Vehicle"])
+
+    # test an Attribute being executed by the Sensemaker
+    symbols: List[SymbolCodeUpdate] = sensemaker.execute(oms_attribute)
+    assert len(symbols) == 0
+
+    # check that symbols exist in Findings table
+    findings = db.execute(
+        select(Finding).filter(Finding.finding_type == FindingType.MIL_SYMBOL_UPDATE.value)).scalars().all()
+
+    assert len(findings) == 0
+
+    # case 7 (Wrong Input Type)
+    oms_relationship = RelationshipRelationship.model_construct(
+        name=SETTINGS.resolution_relationship_name,
+        startNodeId=oms_node.id,
+    )
+
+    symbols: List[SymbolCodeUpdate] = sensemaker.execute(oms_relationship)
+    assert len(symbols) == 0
+
+    # check that symbols do not exist in Findings table
+    findings = db.execute(
+        select(Finding).filter(Finding.finding_type == FindingType.MIL_SYMBOL_UPDATE.value)).scalars().all()
+
+    assert len(findings) == 0
+
+    # ensure findings cleared at end of test cases
     findings = db.execute(delete(Finding))